LiquidSlder-Plugin
A responsive tab plugin that work's both for the WebApp and MobileApp.
Based on the logic of CSS3 style and AngularJs Single Page Application,
it selects a default page as well as shifts to next page without making
any synchrounous call. Implemnting Ajax Logic and BootStrap Design, it
is simple Application with beautiful and easy look responsive plugin.